Add priorDayLimit and priorEntryLimit (#5)
This change adds two new config options: `priorDayLimit` and `priorEntryLimit`. Each of these provide a way to limit the number of menu items display from days _before_ today.

The day limiter simply caps the number of days before today that are shown. For example, if you set it to 1 it will only show entries from the previous day.

The entry limiter counts the number of entries before today's, regardless of which day they are on. The count does respect the configurable sort order (counts in reverse).

You can combine these two options. In order to show a previous entry it has to be within both parameters, not just one.

Closes #4